+++ Model/Ui/Cash/ConfigProvider.php
@@ -44,7 +44,7 @@ class ConfigProvider implements ConfigProviderInterface
     /**
      * Get config
      *
-     * @return \array[][]
+     * @return array
      */
     public function getConfig()
     {